feat(AppProvider): refactor AppProvider import paths and create new AppProvider component to enhance structure and theming capabilities in the application
