+++ apps/client/.storybook/decorators/AppDecorator.tsx
@@ -1,6 +1,6 @@
 import {ReactRenderer} from '@storybook/react';
 import {DecoratorFunction} from '@storybook/types';
-import {AppProvider} from '../../src/components/AppProvider';
+import {AppProvider} from '../../src/components/app/AppProvider';
 
 export const AppDecorator =
     // eslint-disable-next-line react/display-name
